refactor: tidy types for init-app utilities
