+++ src/init-app.utils.ts
@@ -1,4 +1,7 @@
 // Pure (no side effects, idempotent) utility functions.
+import escapeHtml from "xss";
+import { OtherEvent, WorkerErrorEvent } from "./init-app.types";
+
 export function isExternal(url) {
   let startsWithProtocol = !!/^http:\/\/|https:\/\//.exec(url);
   if (!startsWithProtocol) {
@@ -25,81 +28,20 @@ export const fullUrlToPath = (fullUrl) => {
   return path;
 };
 
-// forked allen kim: https://stackoverflow.com/a/47614491/5129731
-export const setInnerHTMLWithScriptsAndOnLoad = async function (elm, html) {
-  elm.innerHTML = html;
-  // throw Error('failure');
-  const scripts = Array.from(elm.querySelectorAll("script"));
-  // console.log('scripts', scripts);
+export function getHtmlFromEvent(eventData: WorkerErrorEvent | OtherEvent) {
+  // TODO... check if ts-pattern is necessary in future
 
-  // A bit hacky but acceptable. Sorting might be better.
-  const scriptsWithDefer = []; // push to end
-  const scriptsWithoutDefer = [];
-  const inlineScripts = [];
-
-  scripts.forEach((script) => {
-    const src = script.getAttribute("src") || "";
-    // Force sql formatter to go to end b/c otherwise codemirror has issues
-    if (src === null || src === "") {
-      inlineScripts.push(script);
-    } else if (script.getAttribute("defer") === null) {
-      scriptsWithoutDefer.push(script);
-    } else {
-      scriptsWithDefer.push(script);
-    }
-  });
-
-  const allRemoteScripts = [...scriptsWithDefer, ...scriptsWithoutDefer];
-
-  // Try to make the scripts wait until the page had loaded before running
-  const asyncLoadPromises = [];
-
-  // insert async scripts in bulk
-  const fragment = new DocumentFragment();
-  allRemoteScripts.forEach((oldScript) => {
-    // console.log('scriptOrder', oldScript.getAttribute('src'));
-    const newScript = document.createElement("script");
-    Array.from(oldScript.attributes).forEach((attr) =>
-      newScript.setAttribute(attr.name, attr.value)
-    );
-    fragment.appendChild(newScript);
-    oldScript.remove();
-    const loadedPromise = new Promise(function (resolve, reject) {
-      newScript.onload = resolve;
-      newScript.onerror = reject;
-    });
-    asyncLoadPromises.push(loadedPromise);
-  });
-
-  // NOTE: this doesn't handle "window.onload" listeners. May need to call that manually
-  const head = document.querySelectorAll("head")[0];
-  head.appendChild(fragment);
-
-  // wait for all scripts to load before executing inline JS
-  // console.log('before')
-  await Promise.all(asyncLoadPromises);
-  // console.log('after');
-
-  // Then insert inline scripts after async items loaded
-  const inlineFragment = new DocumentFragment();
-  inlineScripts.forEach((oldScript) => {
-    const newScript = document.createElement("script");
-    Array.from(oldScript.attributes).forEach((attr) =>
-      newScript.setAttribute(attr.name, attr.value)
-    );
-    newScript.appendChild(document.createTextNode(oldScript.innerHTML));
-    inlineFragment.appendChild(newScript);
-  });
-
-  head.appendChild(inlineFragment);
-
-  // Trigger onloads to finish scripting since this function is async
-  if (window.onload) {
-    console.log("re-dispatching onload");
-    window.onload(); // used by host page
+  if (eventData.type === "error") {
+    return `<div style="padding: 0.5em"><h3>Error</h3><pre>${escapeHtml(
+      eventData.error
+    )}</pre></div>`;
+  } else if (/^text\/html/.exec(eventData.contentType)) {
+    return eventData.text;
+  } else if (/^application\/json/.exec(eventData.contentType)) {
+    return `<pre style="padding: 0.5em">${escapeHtml(
+      JSON.stringify(JSON.parse(eventData.text), null, 4)
+    )}</pre>`;
   }
-  console.log("dispatching scripts loaded");
-  // plugins need to know when all APIS are ready...
-  // window is no good... need to use document.
-  document.dispatchEvent(new CustomEvent("DatasetteLiteScriptsLoaded")); // used by vega
-};
+
+  return `<pre style="padding: 0.5em">${escapeHtml(eventData.text)}</pre>`;
+}
